requirements.txt
Telethon
python-dotenv
asyncio
TgCrypto
